// bidly/apps/server/src/app/open-api/pre-standard/types/pre-standard-thing.type.ts
/**
 * 사전규격 물품
 * @author 최강훈 <ganghun@lepisode.team>
 */
export type PreStandard_Thing = {
  bsnsDivNm: string;
  refNo: string;
  prdctClsfcNoNm: string;
  orderInsttNm: string;
  rlDminsttNm: string;
  asignBdgtAmt: string;
  rcptDt: Date;
  opninRgstClseDt: Date;
  ofclTelNo: string;
  ofclNm: string;
  swBizObjYn: string;
  dlvrTmlmtDt: string;
  dlvrDaynum: string;
  bfSpecRgstNo: string;
  specDocFileUrl1: string;
  specDocFileUrl2: string;
  specDocFileUrl3: string;
  specDocFileUrl4: string;
  specDocFileUrl5: string;
  prdctDtlList: string;
  rgstDt: Date;
  chgDt: string;
  bidNtceNoList: string;
};